+++ includes/Hooks.php
@@ -77,6 +77,7 @@ class Hooks implements BeforePageDisplayHook, GetDoubleUnderscoreIDsHook {
 			'wgPageReaderVoicePitch' => $effective['voicePitch'],
 			'wgPageReaderVoiceRate' => $effective['voiceRate'],
 			'wgPageReaderVoiceGender' => $effective['voiceGender'],
+			'wgPageReaderHighlightEnabled' => $effective['highlightEnabled'],
 			'wgPageReaderPreferredVoices' => $effective['preferredVoices'],
 		] );
 	}